+++ display/util.go
@@ -5,8 +5,10 @@ import (
 	"net/url"
 	"strings"
 
+	"github.com/makeworld-the-better-one/go-gemini"
 	"github.com/spf13/viper"
 	"gitlab.com/tslocum/cview"
+	"golang.org/x/text/unicode/norm"
 )
 
 // This file contains funcs that are small, self-contained utilities.
@@ -73,15 +75,23 @@ func resolveRelLink(t *tab, prev, next string) (string, error) {
 // Example: gemini://gus.guru:1965/ and //gus.guru/.
 // This function will take both output the same URL each time.
 //
+// It will also percent-encode invalid characters, and decode chars
+// that don't need to be encoded. It will also apply Unicode NFC
+// normalization.
+//
 // The string passed must already be confirmed to be a URL.
 // Detection of a search string vs. a URL must happen elsewhere.
 //
 // It only works with absolute URLs.
 func normalizeURL(u string) string {
-	parsed, err := url.Parse(u)
+	u = norm.NFC.String(u)
+
+	tmp, err := gemini.GetPunycodeURL(u)
 	if err != nil {
 		return u
 	}
+	u = tmp
+	parsed, _ := url.Parse(u)
 
 	if parsed.Scheme == "" {
 		// Always add scheme
@@ -102,7 +112,32 @@ func normalizeURL(u string) string {
 	// gemini://example.com -> gemini://example.com/
 	if parsed.Path == "" {
 		parsed.Path = "/"
+	} else {
+		// Decode and re-encode path
+		// This removes needless encoding, like that of ASCII chars
+		// And encodes anything that wasn't but should've been
+		parsed.RawPath = strings.ReplaceAll(url.PathEscape(parsed.Path), "%2F", "/")
+	}
+
+	// Do the same to the query string
+	un, err := gemini.QueryUnescape(parsed.RawQuery)
+	if err == nil {
+		parsed.RawQuery = gemini.QueryEscape(un)
 	}
 
 	return parsed.String()
+}
+
+// fixUserURL will take a user-typed URL and add a gemini scheme to it if
+// necessary. It is not the same as normalizeURL, and that func should still
+// be used, afterward.
+//
+// For example "example.com" will become "gemini://example.com", but
+// "//example.com" will be left untouched.
+func fixUserURL(u string) string {
+	if !strings.HasPrefix(u, "//") && !strings.HasPrefix(u, "gemini://") && !strings.Contains(u, "://") {
+		// Assume it's a Gemini URL
+		u = "gemini://" + u
+	}
+	return u
 }